Address code review feedback.
